# Bug 1182194 - Tweak in order to use source-map.js instead of SourceMap.jsm. r=fitzgen