--- a/build/prefix-utils.jsm
+++ b/build/prefix-utils.jsm
@@ -12,7 +12,8 @@
  * https://github.com/mozilla/source-map/
  */
 
-Components.utils.import('resource://gre/modules/devtools/Require.jsm');
-Components.utils.import('resource://gre/modules/devtools/SourceMap.jsm');
+let loader = Components.classes["@mozilla.org/moz/jssubscript-loader;1"]
+  .getService(Components.interfaces.mozIJSSubScriptLoader);
+loader.loadSubScript("resource://gre/modules/devtools/sourcemap/source-map.js", this);
 
 this.EXPORTED_SYMBOLS = [ "define", "runSourceMapTests" ];