Two small changes.
I'm getting rid of Require.jsm usage on mozilla-central. Instead I'm using builtin module loading shipped within source-map.js and its module loader doesn't take care of `define` second argument `['exports`]`, it just always pass `require` and `exports`.

More context in the bug https://bugzilla.mozilla.org/show_bug.cgi?id=935366