@fitzgen 2015-07-29
LGTM, thanks!

Can you verify that the browser build still works with your changes? `node Makefile.dryice.js` and then run the benchmark. https://github.com/mozilla/source-map/blob/master/bench/README.md

Feel free to merge it in after fixing up that one nit and verifying the browser build still works.