@ochameau 2015-07-30
Benchmark with this patch:
```
Parsing source map
[Stats samples: 8, total: 24379 ms, mean: 3047.375 ms, stddev: 162.02790135136146 ms]

Serializing source map
[Stats samples: 2, total: 26248 ms, mean: 13124 ms, stddev: 86 ms]
```

Without this patch:
```
Parsing source map
[Stats samples: 8, total: 24527 ms, mean: 3065.875 ms, stddev: 139.97835109583983 ms]

Serializing source map
[Stats samples: 2, total: 27254 ms, mean: 13627 ms, stddev: 416 ms]
```

Try run:
  https://treeherder.mozilla.org/#/jobs?repo=try&revision=56770ce87784