@fitzgen 2015-08-04
Thanks. Note that Github doesn't send emails unless you comment, so I had no idea you had fixed this PR up.